pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pulseforge"
version = "0.1.0"
description = "Modular async signal processing engine — audio to TUI visualizer and hardware control"
requires-python = ">=3.11,<3.15"
dependencies = [
    "textual",
    "numpy",
    "scipy",
    "pygame-ce",
]

[project.scripts]
pulseforge = "main:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "inputs", "outputs"]

[tool.hatch.build.force-include]
"main.py" = "main.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

